rocksdb: rename implementation files
