+++ backend/balboa-rocksdb/Makefile
@@ -38,9 +38,9 @@ src-balboa-rocksdb=trace.c daemon.c engine.c mpack.c
 src-balboa-rocksdb-y=$(addprefix ../lib/,$(src-balboa-rocksdb))
 
 hdr-balboa-rocksdb=protocol.h engine.h trace.h daemon.h mpack.h
-hdr-balboa-rocksdb-y=$(addprefix ../lib/,$(hdr-balboa-rocksdb)) rocksdb.h
+hdr-balboa-rocksdb-y=$(addprefix ../lib/,$(hdr-balboa-rocksdb)) rocksdb-impl.h
 
-src-balboa-rocksdb-y+=rocksdb.c main.c
+src-balboa-rocksdb-y+=rocksdb-impl.c main.c
 ifdef D
 target-balboa-rocksdb-y=$(OUT)$(CROSS_PREFIX)balboa-rocksdb-debug
 else
